+++ Sources/HIGThemesContract/HIGTheme+Defaults.swift
@@ -71,6 +71,7 @@ extension HIGTheme {
     public var adminShell: any HIGAdminShellTokens { HIGSystemAdminShellTokens() }
     public var codeBlock: any HIGCodeBlockTokens { HIGSystemCodeBlockTokens() }
     public var carousel: any HIGCarouselTokens { HIGSystemCarouselTokens() }
+    public var mediaRow: any HIGMediaRowTokens { HIGSystemMediaRowTokens() }
     public var photoPicker: any HIGPhotoPickerTokens { HIGSystemPhotoPickerTokens() }
     public var photoEditor: any HIGPhotoEditorTokens { HIGSystemPhotoEditorTokens() }
     public var longTextEditor: any HIGLongTextEditorTokens { HIGSystemLongTextEditorTokens() }