+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -69,6 +69,7 @@ public protocol HIGTheme: Sendable {
     var adminShell: any HIGAdminShellTokens { get }
     var codeBlock: any HIGCodeBlockTokens { get }
     var carousel: any HIGCarouselTokens { get }
+    var mediaRow: any HIGMediaRowTokens { get }
     var photoPicker: any HIGPhotoPickerTokens { get }
     var photoEditor: any HIGPhotoEditorTokens { get }
     var longTextEditor: any HIGLongTextEditorTokens { get }